pyproject.toml
[project]
name = "wandern"
version = "0.0.1"
authors = [{ name = "Shiladitya Bose", email = "shiladitya_basu@live.com" }]
description = "Wandern is a small, no-nonsense database migration tool for python"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "jinja2>=3.1.6",
    "networkx>=3.2.1",
    "pydantic>=2.11.7",
    "questionary>=2.1.0",
    "rich>=13.9.4",
    "typer>=0.13.0",
]

[project.scripts]
wandern = "wandern.__main__:app"
wd = "wandern.__main__:app"

[project.urls]
Homepage = "https://github.com/s-bose/wandern"
Issues = "https://github.com/s-bose/wandern/issues"

[project.optional-dependencies]
postgresql = ["psycopg[binary]>=3.2.9"]
openai = ["pydantic-ai-slim[openai]>=0.7.4"]
google-genai = ["pydantic-ai-slim[google]>=0.7.4"]

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[[tool.mypy.overrides]]
module = ["graphviz"]
ignore_missing_imports = true
check_untyped_defs = true

[tool.uv]
dev-dependencies = ["pytest>=8.3.3", "pytest-asyncio>=0.24.0"]

[dependency-groups]
dev = ["ruff>=0.12.7", "types-networkx>=3.4.2.20250509"]
test = ["pytest-cov>=6.2.1", "testcontainers[postgres]>=4.12.0"]
